+++ Sources/Sentry/include/SentryHub+Private.h
@@ -1,5 +1,4 @@
 #import "SentryHub.h"
-#import "SentryTransactionContext+Private.h"
 
 @class SentryEnvelopeItem, SentryId, SentryScope, SentryTransaction, SentryDispatchQueueWrapper,
     SentryTracer;
